+++ intellij-plugin/src/main/kotlin/com/apollographql/ijplugin/project/ApolloProjectManagerListener.kt
@@ -1,6 +1,5 @@
-package com.apollographql.ijplugin.listeners
+package com.apollographql.ijplugin.project
 
-import com.apollographql.ijplugin.apollo.apolloProjectService
 import com.apollographql.ijplugin.codegen.ApolloCodegenService
 import com.apollographql.ijplugin.util.logd
 import com.intellij.openapi.project.Project
